+++ tests/all_tests/t_f32x4.rs
@@ -995,6 +995,24 @@ fn impl_f32x4_pow_n() {
 }
 
 #[test]
+fn impl_f32x4_is_sign_positive() {
+  let a = f32x4::new([2401.0, -123.0, f32::INFINITY, f32::NEG_INFINITY]);
+  let expected = f32x4::new([f32::from_bits(!0), 0.0, f32::from_bits(!0), 0.0]);
+  let actual = a.is_sign_positive();
+  // Use bitwise equality to accept NaNs as equal.
+  assert_eq!(expected ^ actual, f32x4::ZERO);
+}
+
+#[test]
+fn impl_f32x4_is_sign_negative() {
+  let a = f32x4::new([2401.0, -123.0, f32::INFINITY, f32::NEG_INFINITY]);
+  let expected = f32x4::new([0.0, f32::from_bits(!0), 0.0, f32::from_bits(!0)]);
+  let actual = a.is_sign_negative();
+  // Use bitwise equality to accept NaNs as equal.
+  assert_eq!(expected ^ actual, f32x4::ZERO);
+}
+
+#[test]
 fn impl_f32x4_reduce_add() {
   let p = f32x4::splat(0.001);
   assert_eq!(p.reduce_add(), 0.004);