+++ tests/all_tests/t_f64x8.rs
@@ -1065,6 +1065,60 @@ fn impl_f64x8_powf_multiple() {
 }
 
 #[test]
+fn impl_f64x8_is_sign_positive() {
+  let a = f64x8::new([
+    2401.0,
+    -123.0,
+    f64::INFINITY,
+    f64::NEG_INFINITY,
+    2401.0,
+    -123.0,
+    f64::INFINITY,
+    f64::NEG_INFINITY,
+  ]);
+  let expected = f64x8::new([
+    f64::from_bits(!0),
+    0.0,
+    f64::from_bits(!0),
+    0.0,
+    f64::from_bits(!0),
+    0.0,
+    f64::from_bits(!0),
+    0.0,
+  ]);
+  let actual = a.is_sign_positive();
+  // Use bitwise equality to accept NaNs as equal.
+  assert_eq!(expected ^ actual, f64x8::ZERO);
+}
+
+#[test]
+fn impl_f64x8_is_sign_negative() {
+  let a = f64x8::new([
+    2401.0,
+    -123.0,
+    f64::INFINITY,
+    f64::NEG_INFINITY,
+    2401.0,
+    -123.0,
+    f64::INFINITY,
+    f64::NEG_INFINITY,
+  ]);
+  let expected = f64x8::new([
+    0.0,
+    f64::from_bits(!0),
+    0.0,
+    f64::from_bits(!0),
+    0.0,
+    f64::from_bits(!0),
+    0.0,
+    f64::from_bits(!0),
+  ]);
+  let actual = a.is_sign_negative();
+  // Use bitwise equality to accept NaNs as equal.
+  assert_eq!(expected ^ actual, f64x8::ZERO);
+}
+
+#[test]
 fn impl_f64x8_reduce_add() {
   let p = f64x8::splat(0.001);
   assert_eq!(p.reduce_add(), 0.008);