+++ changelog.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 * Added more float functions.
-
 * Implemented `Rem` for float types.
+* Renamed float function `sign_bit` to `is_sign_negative` and added
+  `is_sign_positive`.
 
 ## 1.3.0